cm: storagestate: move storagestate to itf directory
This patch moves the storagestate interface to the itf directory
and updates the interface.
