+++ src/core/cm/storagestate/tests/CMakeLists.txt
@@ -4,24 +4,30 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-# ######################################################################################################################
-# Target name
-# ######################################################################################################################
-
-set(TARGET_NAME storagestate)
+set(TARGET_NAME storagestate_test)
 
 # ######################################################################################################################
-# Headers
+# Sources
 # ######################################################################################################################
 
-set(HEADERS itf/sender.hpp itf/state.hpp itf/storage.hpp itf/storagestate.hpp)
+set(SOURCES storagestate.cpp)
+
+# ######################################################################################################################
+# Libraries
+# ######################################################################################################################
+
+set(LIBRARIES aos::core::cm::storagestate GTest::gmock_main)
 
 # ######################################################################################################################
 # Target
 # ######################################################################################################################
 
-add_module(TARGET_NAME ${TARGET_NAME} HEADERS ${HEADERS})
-
-if(WITH_TEST)
-    add_subdirectory(tests)
-endif()
+add_test(
+    TARGET_NAME
+    ${TARGET_NAME}
+    LOG_MODULE
+    SOURCES
+    ${SOURCES}
+    LIBRARIES
+    ${LIBRARIES}
+)
