+++ src/core/cm/storagestate/tests/storagestate.cpp
@@ -1,0 +1,9 @@
+/*
+ * Copyright (C) 2025 EPAM Systems, Inc.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <core/cm/storagestate/itf/sender.hpp>
+#include <core/cm/storagestate/itf/storage.hpp>
+#include <core/cm/storagestate/itf/storagestate.hpp>
